+++ src/main/kotlin/com/nylas/resources/RedirectUris.kt
@@ -14,64 +14,74 @@ import com.nylas.util.JsonHelper
 class RedirectUris(client: NylasClient) : Resource<RedirectUri>(client, RedirectUri::class.java) {
   /**
    * Return all Redirect URIs
+   * @param overrides Optional request overrides to apply
    * @return The list of Redirect URIs
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
-  fun list(): ListResponse<RedirectUri> {
+  @JvmOverloads
+  fun list(overrides: RequestOverrides? = null): ListResponse<RedirectUri> {
     val path = "v3/applications/redirect-uris"
-    return listResource(path)
+    return listResource(path, overrides = overrides)
   }
 
   /**
    * Return a Redirect URI
    * @param redirectUriId The id of the Redirect URI to retrieve.
+   * @param overrides Optional request overrides to apply
    * @return The Redirect URI
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
-  fun find(redirectUriId: String): Response<RedirectUri> {
+  @JvmOverloads
+  fun find(redirectUriId: String, overrides: RequestOverrides? = null): Response<RedirectUri> {
     val path = String.format("v3/applications/redirect-uris/%s", redirectUriId)
-    return findResource(path)
+    return findResource(path, overrides = overrides)
   }
 
   /**
    * Create a Redirect URI
    * @param requestBody The values to create the Redirect URI with
+   * @param overrides Optional request overrides to apply
    * @return The created Redirect URI
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
-  fun create(requestBody: CreateRedirectUriRequest): Response<RedirectUri> {
+  @JvmOverloads
+  fun create(requestBody: CreateRedirectUriRequest, overrides: RequestOverrides? = null): Response<RedirectUri> {
     val path = "v3/applications/redirect-uris"
     val serializedRequestBody = JsonHelper.moshi()
       .adapter(CreateRedirectUriRequest::class.java)
       .toJson(requestBody)
 
-    return createResource(path, serializedRequestBody)
+    return createResource(path, serializedRequestBody, overrides = overrides)
   }
 
   /**
    * Update a Redirect URI
    * @param redirectUriId The id of the Redirect URI to update.
    * @param requestBody The values to update the Redirect URI with
+   * @param overrides Optional request overrides to apply
    * @return The updated Redirect URI
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
-  fun update(redirectUriId: String, requestBody: UpdateRedirectUriRequest): Response<RedirectUri> {
+  @JvmOverloads
+  fun update(redirectUriId: String, requestBody: UpdateRedirectUriRequest, overrides: RequestOverrides? = null): Response<RedirectUri> {
     val path = String.format("v3/applications/redirect-uris/%s", redirectUriId)
     val serializedRequestBody = JsonHelper.moshi()
       .adapter(UpdateRedirectUriRequest::class.java)
       .toJson(requestBody)
 
-    return patchResource(path, serializedRequestBody)
+    return patchResource(path, serializedRequestBody, overrides = overrides)
   }
 
   /**
    * Delete a Redirect URI
    * @param redirectUriId The id of the Redirect URI to delete.
+   * @param overrides Optional request overrides to apply
    * @return The deleted Redirect URI
    */
   @Throws(NylasApiError::class, NylasSdkTimeoutError::class)
-  fun destroy(redirectUriId: String): DeleteResponse {
+  @JvmOverloads
+  fun destroy(redirectUriId: String, overrides: RequestOverrides? = null): DeleteResponse {
     val path = String.format("v3/applications/redirect-uris/%s", redirectUriId)
-    return destroyResource(path)
+    return destroyResource(path, overrides = overrides)
   }
 }