+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Nylas Java SDK Changelog
 
+### [Unreleased]
+
+### Added
+* Added support for custom headers field for Drafts and Messages
+* Added support for overriding various fields of outgoing requests
+
 ## [2.2.1] - Released 2024-03-05
+
+### Added
+* Added support for `provider` field in code exchange response
 
 ### Changed
 * Improved message sending and draft create/update performance